requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep learning
torch>=1.10.0
helical>=0.1.0

# Data handling
anndata>=0.8.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Performance profiling
psutil>=5.8.0
GPUtil>=1.4.0
memory-profiler>=0.60.0

# Optimization
optuna>=3.0.0
plotly>=5.0.0  # Optional, for Optuna visualizations

# Optional: ONNX and TensorRT
onnx>=1.12.0
onnxruntime-gpu>=1.12.0
tensorrt>=8.0.0  # Install from NVIDIA

# Optional: Weights & Biases for tracking
wandb>=0.12.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0





